tests/wpt/tests/css/css-fonts/parsing/font-variant-caps-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: getComputedStyle().fontVariantCaps</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-caps-prop">
<meta name="assert" content="font-variant-caps computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-variant-caps', 'normal');
test_computed_value('font-variant-caps', 'small-caps');
test_computed_value('font-variant-caps', 'all-small-caps');
test_computed_value('font-variant-caps', 'petite-caps');
test_computed_value('font-variant-caps', 'all-petite-caps');
test_computed_value('font-variant-caps', 'unicase');
test_computed_value('font-variant-caps', 'titling-caps');
</script>
</body>
</html>
